[Fix] reload 4번 호출되는 문제 해결, 오타 수정, 오류 모두 해결했습니다
