+++ AlarmButler/WorldClock/Views/TimeZoneTVCell.swift
@@ -8,8 +8,7 @@
 import UIKit
 import SwiftUI
 
-class TimeZoneTableViewCell: UITableViewCell {
-    
+class TimeZoneTableViewCell: UITableViewCell {//일본
     static let identi = "AddTableViewCell"
     
     var data: String? {
@@ -39,9 +38,11 @@ class TimeZoneTableViewCell: UITableViewCell {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
+    
     override func awakeFromNib() {
         super.awakeFromNib()
     }
+    
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)
     }
