+++ AlarmButler/WorldClock/Views/WorldClockVC.swift
@@ -9,24 +9,30 @@ import Foundation
 import UIKit
 import SnapKit
 
-class WorldClockViewController: UIViewController {
-    
+class WorldClockViewController: UIViewController {    
     let clockDataManager = WorldClockManager.shared
     
     // 세계 시계 데이터 배열
-    var colckData: [WorldClockEntity] {
+    var clockData: [WorldClockEntity] {
+        ///0번째에 세계시계 라벨
+        ///1번째에 셰계시간 데이터
         get {
             return clockDataManager.getSavedWorldClock()
         }
     }
     
+    private lazy var worldLabel = {
+        let worldLabel = UILabel()
+        worldLabel.text = "세계 시계"
+        worldLabel.textColor = .black
+        worldLabel.font = UIFont.boldSystemFont(ofSize: 40)
+        return worldLabel
+    }()
+    
     private lazy var tableView = {
         let tableView = UITableView()
-        tableView.register(WorldClockViewCell.self, forCellReuseIdentifier: WorldClockViewCell.identi)
         tableView.register(CustomTableViewCell.self, forCellReuseIdentifier: CustomTableViewCell.identi)
-        
         tableView.rowHeight = UITableView.automaticDimension
-            tableView.estimatedRowHeight = 44
         return tableView
     }()
     
@@ -60,83 +66,64 @@ class WorldClockViewController: UIViewController {
     }
 }
 
-
 extension WorldClockViewController {
     private func addSubView(){
         view.addSubview(navigationBar)
         view.addSubview(tableView)
+        view.addSubview(worldLabel)
     }
-    
-    
+        
     private func autoLayout(){
         navigationBar.snp.makeConstraints { make in
             make.top.leading.trailing.equalTo(view.safeAreaLayoutGuide)
             make.height.equalTo(44)
         }
-        tableView.snp.makeConstraints { make in
+        worldLabel.snp.makeConstraints { make in
             make.top.equalTo(navigationBar.snp.bottom)
+            make.leading.equalToSuperview().offset(8)
+            make.height.equalTo(60)
+        }
+        tableView.snp.makeConstraints { make in
+            make.top.equalTo(worldLabel.snp.bottom)
             make.leading.trailing.bottom.equalTo(view.safeAreaLayoutGuide)
         }
     }
-    
-    
+        
     // TimeZoneViewController로부터 전달받은 데이터를 처리하는 함수
-    func addClockData(_ region: String, _ timeZone: TimeZone) {
-        clockDataManager.saveWorldClockData(newRegion: region, newTimeZone: timeZone) {
-            // 저장이 완료되면 화면을 갱신. (예: 테이블 뷰 리로드)
-            self.tableView.reloadData()
-        }
-    }
+//    func addClockData(_ region: String, _ timeZone: TimeZone) {
+//        print("addClockData")
+//        clockDataManager.saveWorldClockData(newRegion: region, newTimeZone: timeZone) {
+//        }
+//    }
 }
-
 
 extension WorldClockViewController: UITableViewDelegate, UITableViewDataSource {
     //셀이 비었을 때 세계시계 없음 Label
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return colckData.count
+        return clockData.count
+        
     }
-    
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        if indexPath.row == 0 {
-            let cell = tableView.dequeueReusableCell(withIdentifier: WorldClockViewCell.identi, for: indexPath) as! WorldClockViewCell
-            
-            return cell
-        } else {
-            let cell = tableView.dequeueReusableCell(withIdentifier: CustomTableViewCell.identi, for: indexPath) as! CustomTableViewCell
-            
-            cell.clockData = colckData[indexPath.row]
-            
-            return cell
-        }
-    }
-    
-    // 테이블 뷰의 특정 행을 편집할 수 있는지 여부 설정
-    func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        guard let cell = tableView.cellForRow(at: indexPath) else {
-            return false // 셀이 없으면 편집할 수 없음
-        }
-        // 만약 셀이 WorldClockViewCell이면 편집할 수 없음
-        if cell is WorldClockViewCell {
-            return false
-        }
-        return true
+        let cell = tableView.dequeueReusableCell(withIdentifier: CustomTableViewCell.identi, for: indexPath) as! CustomTableViewCell
+        cell.clockData = clockData[indexPath.row]
+        return cell
+        
     }
     
     // 테이블 뷰의 특정 행을 편집할 때의 동작 설정
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
         if(editingStyle == .delete){
-            
-            clockDataManager.removeClock(deleteTarget: colckData[indexPath.row]) {
+            clockDataManager.removeClock(deleteTarget: clockData[indexPath.row]) {
             tableView.reloadData()
             }
         }
     }
 }
 
-
 extension WorldClockViewController {
     @objc private func tappedButton() {
+        print("tappedButton")
         //let firstVC = WorldClockViewController()
         let VC = TimeZoneController()
         VC.delegate = self
@@ -146,7 +133,6 @@ extension WorldClockViewController {
         //firstVC.tableView.reloadData()
     }
 }
-
 
 extension WorldClockViewController: TimeZoneViewControllerDelegate {
     func didUpdateWorldClockData() {
